old APU changes; format; lint

+++ src/gb-lib/ppu/layer/tilemap.cpp
@@ -16,7 +16,10 @@ auto TileMap::getTile(const TileAddress& address) -> Tile { return tiles_->getTi
 
 auto TileMap::toFlatAddress(const TileAddress& address) -> uint8_t { return address.x * TileMapSize + address.y; }
 
-auto TileMap::getIndex(uint8_t flatAddress) const -> uint8_t { return mem_->getByte(baseAdress() + flatAddress).get(); }
+auto TileMap::getIndex(uint8_t flatAddress) const -> int8_t
+{
+    return static_cast<int8_t>(mem_->getByte(baseAdress() + flatAddress).get());
+}
 
 auto TileMap::baseAdress() const -> address_type
 {
